+++ app/src/app/AppLayout.tsx
@@ -3,6 +3,7 @@ import { Inter } from 'next/font/google'
 import './globals.css'
 import Header from '@/components/Header'
 import { useRouter } from "next/navigation";
+import { useState, useEffect } from 'react';
 
 const inter = Inter({ subsets: ['latin'] })
 export default function RootLayout({
@@ -10,13 +11,34 @@ export default function RootLayout({
 }: {
   children: React.ReactNode
 }) {
+  const [isPcType, setIsPcType] = useState(false)
+  useEffect(() => {
+    const mobileTypeList = ["iPhone", "iPad", "iPod", "Android"];
+    const machineType = navigator.userAgent;
+    const mobileTypeCheck = mobileTypeList.filter((item) => {
+      return machineType.search(item) != -1
+    })
+    if(mobileTypeCheck.length > 0){
+      setIsPcType(false)
+    }else{
+      setIsPcType(true)
+    }
+    console.log(machineType)
+  })
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Header/>
-        <div className='mt-6 ml-28 mr-28 mb-6'>
-          {children}
-        </div>
+        {
+          isPcType?
+          <div>
+            <Header/>
+            <div className='mt-6 ml-28 mr-28 mb-6'>
+              {children}
+            </div>
+          </div>:
+          <div>a</div>
+        }
+        
       </body>
     </html>
   )
